should use plain int for index, since i may be printed using %d; close #2127

+++ src/codegen.c
@@ -2627,7 +2627,7 @@ static void
 codedump(mrb_state *mrb, mrb_irep *irep)
 {
 #ifdef ENABLE_STDIO
-  size_t i;
+  int i;
   int ai;
   mrb_code c;
 
@@ -2635,7 +2635,7 @@ codedump(mrb_state *mrb, mrb_irep *irep)
   printf("irep %p nregs=%d nlocals=%d pools=%d syms=%d reps=%d\n", irep,
          irep->nregs, irep->nlocals, (int)irep->plen, (int)irep->slen, (int)irep->rlen);
 
-  for (i = 0; i < irep->ilen; i++) {
+  for (i = 0; i < (int)irep->ilen; i++) {
     ai = mrb_gc_arena_save(mrb);
     printf("%03d ", i);
     c = irep->iseq[i];
